Feature/Franco-BP/Se remodela el id con Integer y se hace logica para BD en Field y Company

+++ src/main/java/com/fulboapp/dao/impl/UserDAOImpl.java
@@ -17,7 +17,7 @@ public class UserDAOImpl implements IUserDAO {
   private IUserRepo repo;
 
   @Override
-  public UserDTO findUser(Long userID) {
+  public UserDTO findUser(Integer userID) {
     return new UserDTO();
   }
